Rename "ConfirmedCasesDetailsCard" to "PatientsBreakdownCard"
